+++ Assets/Scripts/World/BlockType.cs
@@ -3,7 +3,7 @@
 namespace World {
     [Serializable]
     public struct BlockType {
-        public int Id;
+        public byte Id;
         public string _Name;
         public bool IsTransparent;
         public int BottomId;
